move column selection actions to columnSelection dir
