+++ src/modules/editorStore/__tests__/tablatureSlice.test.ts
@@ -5,12 +5,12 @@ import { act, cleanup, renderHook } from '@testing-library/react';
 import { changeInstrument } from '../actions/changeInstrument';
 import { changeTuning } from '../actions/changeTuning';
 import { clearSelectedColumns } from '../actions/clearSelectedColumns';
+import { setColumnSelection } from '../actions/columnSelection/setColumnSelection';
 import { duplicateSelectedColumns } from '../actions/duplicateSelectedColumns';
 import { insertColumnsAtSelection } from '../actions/insertColumnsAtSelection';
 import { pushBlankSection } from '../actions/pushBlankSection';
 import { resetStore } from '../actions/resetStore';
 import { resetTablature } from '../actions/resetTablature';
-import { setColumnSelection } from '../actions/setColumnSelection';
 import { setSelectedColumnsCellModifiers } from '../actions/setSelectedColumnsCellModifiers';
 import { setSelectedColumnsFret } from '../actions/setSelectedColumnsFret';
 import { CELL_MODIFIERS, electricBass } from '../constants';
