+++ src/modules/editorStore/actions/columnSelection/columnSelectionStart.ts
@@ -1,6 +1,6 @@
-import { BLANK_SELECTION } from '../constants';
-import { useTablatureEditorStore } from '../useTablatureEditorStore';
-import { validateColumnSelection } from '../utils/validateColumnSelection';
+import { BLANK_SELECTION } from '../../constants';
+import { useTablatureEditorStore } from '../../useTablatureEditorStore';
+import { validateColumnSelection } from '../../utils/validateColumnSelection';
 
 export const columnSelectionStart = (sectionIndex: number, columnIndex: number) =>
 	useTablatureEditorStore.setState((state) => {
